Conflicts:
	src/main/java/demo/application/service/UserService.java

+++ src/main/java/demo/application/domain/Role.java
@@ -23,7 +23,6 @@ import lombok.NoArgsConstructor;
 public class Role {
     @Id 
     @GeneratedValue(strategy=AUTO)
-    private Long id;
-    
+    private Integer id;
     private String name;
 }
